Clean up non-parameterized tests in the wrong place (#131049)
In the course of other work I found a few places where we were creating non-parameterized tests from within the parameterized test drivers for a few ESQL functions. This causes those tests to be run for every parameter combination, even though the tests themselves do not change anything, resulting in a lot of extra test overhead for no additional coverage.

I cleaned up three classes:

    EndsWithTests ran 832 tests before this, and now runs 640 tests
    StartsWithTests ran 208 before this, and now runs 160
    InTests ran 168 before this, and now runs 28

So overall, that's 369 redundant test runs removed (including the fact that the 11 tests I moved still run once in their new classes), and further savings if we later expand those parameterized tests.
